start templating demo

+++ lecture-18_user_roles/demo-1-user-roles-finish/routes/apiv1.js
@@ -5,10 +5,14 @@ var router = express.Router();
 router.get('/users/me', (req, res) => {
     let session = req.session
     if(session.isAuthenticated){
+      let roles = session.account.idTokenClaims.roles ? session.account.idTokenClaims.roles : []
+      let isModerator = roles.includes("Content.Moderator")
+        
       res.json({
           status: "logged in",
           name: session.account.name,
-          username: session.account.username
+          username: session.account.username,
+          isModerator: isModerator
       });
     } else {
       res.json({status: "not logged in (anonymous)"})
@@ -35,9 +39,15 @@ router.post('/comments', async (req, res) => {
 
 router.delete("/comments", async (req, res, next) =>{
     if(req.session.isAuthenticated){
-        let commentId = req.body.commentId
-        await req.models.Comment.deleteOne({_id: commentId})
-        res.json({status: "success"})
+        let roles = req.session.account.idTokenClaims.roles ? req.session.account.idTokenClaims.roles : []
+        let isModerator = roles.includes("Content.Moderator")
+        if(isModerator){
+            let commentId = req.body.commentId
+            await req.models.Comment.deleteOne({_id: commentId})
+            res.json({status: "success"})
+        } else {
+            res.json({status: "Error: Not a moderator"})
+        }
     } else {
         res.json({status: "Error: Not logged in"})
     }
